Updated to use python agent and flask app

+++ demo.sh
@@ -8,13 +8,16 @@ cleanup() {
 
 trap cleanup SIGINT
 
-export APPOPTICS_API_TOKEN=change-me
+if [ -z "$APPOPTICS_API_TOKEN" ]; then
+    echo please set the APPOPTICS_API_TOKEN environment variable
+    exit 1
+fi
 docker-compose down -v --remove-orphans
 docker-compose up -d
 
 sleep 2
 
 while true; do
-    curl -i http://localhost:8001/chain?target=http://php-logdemo/logging.php
+    curl -i http://localhost:8001/rpc/http://php-logdemo/logging.php
     sleep 1
 done
